move UserNameAndLogo to components, update imports
