+++ src/features/posts/components/PostExcerpt/PostExcerpt.js
@@ -4,7 +4,6 @@ import React, { useCallback, useState } from 'react'
 import { useSelector, useDispatch } from 'react-redux'
 import { Link } from 'react-router-dom'
 import { CSSTransition } from 'react-transition-group'
-import { UserNameAndLogo } from '../../../users/UserNameAndLogo'
 import { TimeAgo } from '../../../../components/TimeAgo'
 import {
   removePost,
