+++ src/features/users/components/UserNameAndLogo/index.js
@@ -1,0 +1,2 @@
+import UserNameAndLogo from './UserNameAndLogo'
+export default UserNameAndLogo
